Helper Functions

These functions were useful to me while developing the main Diff functionality. Hopefully they are useful to anyone using the package. They are presented here in no particular order.

CurricularAnalyticsDiff.prereq_print — Function

prereq_print(prereqs::Set{AbstractString})

Return a string representing the names of all courses in the prereqs set.

Usually only used for prereqs, hence the name. It can be used with any normal set of courses.

CurricularAnalyticsDiff.get_course_prereqs — Function

get_course_prereqs(course::Course, curriculum::Curriculum)

Return a vector containing the courses that are prerequisites of course in curriculum.

CurricularAnalyticsDiff.course_from_name — Function

course_from_name(course_name::AbstractString, curriculum::Curriculum)

Return the course object with the name course name in the curriculum. In newer versions (since 0.1.5) this checks for a match with course prefix with course number following the UCSD format of "prefix num"

Serves as a human-readable alternative to course_from_id in the main Curricular Analytics package.

CurricularAnalyticsDiff.pretty_print_course_names — Function

pretty_print_course_names(courses::Vector{})

Print course names prettily to console.

courses should be a vector of course names for this to work correctly.

CurricularAnalyticsDiff.courses_to_course_names — Function

courses_to_course_names(courses::Vector{})

Return an array of the course names corresponding to the given course objects in courses (should be a vector of course objects).

CurricularAnalyticsDiff.courses_that_depend_on_me — Function

courses_that_depend_on_me(course_me::Course, curriculum::Curriculum)

Return an array of courses that represent the first level of course_me's unblocked field in curriculum (as defined by Curricular Analytics).

That is the list of courses that explicitly list course_me as a prerequisite. Includes co-requisites.

CurricularAnalyticsDiff.blocking_factor_investigator — Function

blocking_factor_investigator(course_me::Course, curriculum::Curriculum)

Return the list of courses that comprise course_me's unblocked field (as defined by Curricular Analytics).

CurricularAnalyticsDiff.delay_factor_investigator — Function

delay_factor_investigator(course_me::Course, curriculum::Curriculum)

Return a list representing a course path in curriculum passing through course_me with length equal to course_me's delay factor. It is not always the same path as the one highlighted in the visualization package.

CurricularAnalyticsDiff.centrality_investigator — Function

centrality_investigator(course_me::Course, curriculum::Curriculum)

Return a list of lists containg the paths that make up the centrality of course_me in curriculum.

Each list is one such path.

CurricularAnalyticsDiff.longest_path_to_me — Function

longest_path_to_me(course_me::Course, curriculum::Curriculum, filter_course::Course, filter::Bool=false)

Returns the longest path in curriculum up to course_me.

If the filter option is enabled, the aforementioned path is one that contains filter_course.

CurricularAnalyticsDiff.snippet — Function

snippet(course::Course, curriculum::Curriculum)

Returns a sub-curriculum of the original including only the courses that compose course's centrality paths