Result Print

This is a group of functions that print the results of a curricular diff operation in a human-legible way. There are two flavors: executive summary and pretty print.

Executive summary is a simplified version with little explanation of why the changes have occurred.

Pretty print is a more detailed version that attempts to explain why changes have occurred.

Executive Summary

CurricularAnalyticsDiff.executive_summary_curriculum — Function

executive_summary_course(curriculum_results::Dict{})

Print a simplified version of the results of curricular diff.

Separates between courses that do and don't have a match.

CurricularAnalyticsDiff.executive_summary_unmatched_course — Function

executive_summary_unmatched_course(results::Dict{}, course_name::AbstractString)

Prints a simplified version of the results for a course with no match and name course_name

Usually, these results are fairly uninteresting because their metrics are zeroed out in one of the two curricula being compared.

CurricularAnalyticsDiff.executive_summary_course — Function

executive_summary_course(results::Dict{String,Any}, course_name::AbstractString)

Print a simplified version of the results of curricular diff for a course with name course_name.

It prints results for centrality, blocking factor and delay factor. Complexity is a linear combination of blocking and delay (and the scale isn't exactly decided yet) so it is left to the user to infer.

Pretty Print

CurricularAnalyticsDiff.pretty_print_curriculum_results — Function

pretty_print_curriculum_results(curriculum_results::Dict{}, desired_stat::DesiredStat)

Prints detailed results of a curricular diff for a whole curriculum and the desired_stat.

DesiredStat is an enum that can take on:

ALL: All the other options at once.

CEN: Centrality

COM: Complexity (Blocking and Delay factors)

BLO: Blocking Factor

DEL: Delay Factor

PRE: Prerequisite Chain

CurricularAnalyticsDiff.pretty_print_course_results — Function

pretty_print_course_results(results::Dict{String,Any}, course_name::AbstractString, desired_stat::DesiredStat)

Print the results of a curricular diff for a course with name course_name for the desired_stat.

DesiredStat is an enum that can take on:

ALL: All the other options at once.

CEN: Centrality

COM: Complexity (Blocking and Delay factors)

BLO: Blocking Factor

DEL: Delay Factor

PRE: Prerequisite Chain

CurricularAnalyticsDiff.pretty_print_prereq_changes — Function

pretty_print_prereq_changes(results::Dict{String,Any})

Print changes in a prereq chain.

CurricularAnalyticsDiff.pretty_print_delay_factor_results — Function

pretty_print_delay_factor_results(results::Dict{String,Any})

Print detailed results of a curricular diff for delay factor only.

It prints the changes in the delay factor and where the courses involved have changed.

CurricularAnalyticsDiff.pretty_print_blocking_factor_results — Function

pretty_print_blocking_factor_results(results::Dict{String,Any})

Print detailed results of a curricular diff for blocking factor only.

It prints the changes in the unblocked fields and where the courses involved have changed.

CurricularAnalyticsDiff.pretty_print_complexity_results — Function

pretty_print_complexity_results(results::Dict{String,Any})

Print the detailed results of a curricular diff for complexity only.

It calls the corresponding blocking and delay factor methods, but it does use the scale-adjusted complexity values.

CurricularAnalyticsDiff.pretty_print_centrality_results — Function

pretty_print_centrality_results(results::Dict{String,Any})

Print the detailed results of a curricular diff for centrality only.

It prints the paths that have changed and why. It can get a little long since there are usually so many centrality paths.