What If

This set of functions simply generate curricula where a course or prerequisite has been added or removed from a base curricula. The intended use is for them to be used with curricular diff to see how small changes can affect a curriculum.

CurricularAnalyticsDiff.add_course — Function

add_course(course_name, curr, credit_hours, prereqs, dependencies)

Return a copy of curr where a new course with the provided information has been added.

Arguments

course_name::AbstractString: The name of the course to add.

curr::Curriculum: The curriculum to add a course to.

credit_hours::Real: How many credit hours the new course is worth.

prereqs::Dict: The names of the prerequisites for the new course and their requisite type.

dependencies::Dict: The names of the courses that would have the new course as a prerequisite and the requisite type.

CurricularAnalyticsDiff.remove_course — Function

remove_course(course_name::AbstractString, curr::Curriculum)

Return a copy of curr where the course with name course_name has been removed.

It is removed from all of the prerequisite chains it was in.

CurricularAnalyticsDiff.add_prereq — Function

add_prereq(course_name::AbstracString, added_prereq::AbstracString, curr::Curriculum, reqtype::Requisite)

Return a copy of curr where the course with name added_prereq has been added as a requisite of type reqtype to the course with name course_name.

CurricularAnalyticsDiff.remove_prereq — Function

remove_prereq(course_name::AbstractString, to_remove::AbstractString, curr::Curriculum)

Return a copy of curr where the course with name to_remove has been removed as a prerequisite of the course with name course_name.