What If Institutional

Whereas What If simply asks how adding/removing a course/prerequisite affects a particular curriculum, these functions attempt to answer the same question for a curriculum that represents all the collected curricula of an institution. The procedure to generate such a file is described elsewhere, as the current build is UCSD-only. The point is that all courses in the special curriculum should have a list of the original curricula they come from in the canonical name field.

CurricularAnalyticsDiff.delete_prerequisite_institutional — Function

delete_prerequisite_institutional(target, prereq, curriculum)

Deletes prereq as a prerequisite to target in curriculum.

CurricularAnalyticsDiff.delete_course_institutional — Function

delete_course_institutional(course_to_remove_name, curriculum)

Removes the course with name matching course to remove name from curriculum and prints how many degree plans were affected.

CurricularAnalyticsDiff.add_course_institutional — Function

add_course_institutional(new_course_name, curriculum, new_course_credit_hours, prereqs, dependencies)

Adds a course with name new course name and provided characteristics to curriculum. Analyzes how many degree plans are affected.

CurricularAnalyticsDiff.add_prereq_institutional — Function

add_prereq_institutional(curriculum, course_with_new_prereq, prereq)

Adds prereq as a prerequisite to course with new prereq.

CurricularAnalyticsDiff.print_affected_plans — Function

print_affected_plans(affected_plans)

Prints a list of the plans affected by a change.