Helper Functions

These functions were useful to me while developing the main Diff functionality. Hopefully they are useful to anyone using the package. They are presented here in no particular order.

CurricularAnalyticsDiff.prereq_print — Function

prereq_print(prereqs)

Returns a string representing the names of all courses in the set. Usually only used for prereqs, hence the name.

CurricularAnalyticsDiff.get_course_prereqs — Function

get_course_prereqs(course,curriculum)

Returns a vector containing the courses that are prerequisites of the given course.

CurricularAnalyticsDiff.course_from_name — Function

course_from_name(course_name, curriculum)

Returns the course object with the given course name in the given curriculum. Serves as a human-readable alternative to course_from_id in the main Curricular Analytics package.

CurricularAnalyticsDiff.pretty_print_course_names — Function

pretty_print_course_names(courses)

Prints course names prettily to console.

CurricularAnalyticsDiff.courses_to_course_names — Function

courses_to_course_names(courses)

Returns a vector of the course names corresponding to the given course objects.

CurricularAnalyticsDiff.courses_that_depend_on_me — Function

courses_that_depend_on_me(course_me, curriculum)

Returns an array of courses that represent the first level of course me's unblocked field (as defined by Curricular Analytics). That is the list of courses that explicitly list course me as a prerequisite. Includes co-requisites.

CurricularAnalyticsDiff.blocking_factor_investigator — Function

blocking_factor_investigator(course_me, curriculum)

Returns the list of courses that comprise course me's unblocked field.

CurricularAnalyticsDiff.delay_factor_investigator — Function

delay_factor_investigator(course_me, curriculum)

Returns a list representing a course path in curriculum passing through course me with length equal to course me's delay factor. It is not always the same path as the one highlighted in the visualization package.

CurricularAnalyticsDiff.centrality_investigator — Function

centrality_investigator(course_me,curriculum)

Returns a list of lists containg the paths that make up the centrality of course me in the curriculum. Each list is one such path.

CurricularAnalyticsDiff.longest_path_to_me — Function

longest_path_to_me(course_me,curriculum,filter_course,filter)

Returns the longest path in curriculum up to course me. If the filter option is enabled, the aforementioned path is one that contains the filter course.