Result Print

This is a group of functions that print the results of a curricular diff operation in a human-legible way. There are two flavors: executive summary and pretty print.

Executive summary is a simplified version with little explanation of why the changes have occurred.

Pretty print is a more detailed version that attempts to explain why changes have occurred.

Executive Summary

CurricularAnalyticsDiff.executive_summary_curriculum — Function

executive_summary_course(curriculum_results)

Prints a simplified version of the results of curricular diff. Separates between courses that do and don't have a match.

CurricularAnalyticsDiff.executive_summary_unmatched_course — Function

executive_summary_unmatched_course(results, course_name)

Prints a simplified version of the results for a course with no match. Usually, these results are fairly uninteresting because their metrics are zeroed out in one of the two curricula being compared.

CurricularAnalyticsDiff.executive_summary_course — Function

executive_summary_course(results, course_name)

Prints a simplified version of the results of curricular diff for a given course. It prints results for centrality, blocking factor and delay factor. Complexity is a linear combination of blocking and delay (and the scale isn't exactly decided yet) so it is left to the user to infer.

Pretty Print

CurricularAnalyticsDiff.pretty_print_curriculum_results — Function

pretty_print_curriculum_results(curriculum_results, desired_stat)

Prints detailed results of a curricular diff for a whole curriculum and a desired statistic.

CurricularAnalyticsDiff.pretty_print_course_results — Function

pretty_print_course_results(results, course_name, desired_stat)

Prints the results of a curricular diff for a given course and given statistic

CurricularAnalyticsDiff.pretty_print_prereq_changes — Function

pretty_print_prereq_changes(results)

Prints changes in a prereq chain.

CurricularAnalyticsDiff.pretty_print_delay_factor_results — Function

pretty_print_delay_factor_results(results)

Prints detailed results of a curricular diff for delay factor. It prints the changes in the delay factor and where the courses involved have changed.

CurricularAnalyticsDiff.pretty_print_blocking_factor_results — Function

pretty_print_blocking_factor_results(results)

Prints detailed results of a curricular diff for blocking factor. It prints the changes in the unblocked fields and where the courses involved have changed.

CurricularAnalyticsDiff.pretty_print_complexity_results — Function

pretty_print_complexity_results(results)

Prints the detailed results of a curricular diff for complexity only. It calls the corresponding blocking and delay factor mathods, but it does use the scale-adjusted complexity values.

CurricularAnalyticsDiff.pretty_print_centrality_results — Function

pretty_print_centrality_results(results)

Prints the detailed results of a curricular diff for centrality only. It prints the paths that have changed and why. It can get a little long since there are usually so many centrality paths.