What If Institutional

Whereas What If simply asks how adding/removing a course/prerequisite affects a particular curriculum, these functions attempt to answer the same question for a curriculum that represents all the collected curricula of an institution. The procedure to generate such a file is described elsewhere, as the current build is UCSD-only. The point is that all courses in the special curriculum should have a list of the original curricula they come from in the canonical name field.

CurricularAnalyticsDiff.delete_prerequisite_institutional — Function

delete_prerequisite_institutional(target::AbstractString, prereq::AbstractString, curriculum::Curriculum)

Remove the course with name prereq from being a prerequisite to the course with name target in curriculum and print how many degree plans were affected.

CurricularAnalyticsDiff.delete_course_institutional — Function

delete_course_institutional(course_to_remove_name::AbstractString, curriculum::Curriculum)

Remove the course with name course_to_remove_name from curriculum and print how many degree plans were affected.

CurricularAnalyticsDiff.add_course_institutional — Function

add_course_institutional(new_course_name::AbstractString, curriculum::Curriculum, new_course_credit_hours::Real, prereqs::Dict, dependencies::Dict)

Add a course with name new_course_name and provided characteristics to curriculum` and print how many degree plans are affected.

CurricularAnalyticsDiff.add_prereq_institutional — Function

add_prereq_institutional(curriculum::Curriculum, course_with_new_prereq::AbstractString, prereq::AbstractString)

Print how many plans are affected by adding a prerequisite to course_with_new_prereq in curriculum.

Note that it currently does not actually add prereq as a prerequisite to course_with_new_prereq

CurricularAnalyticsDiff.print_affected_plans — Function

print_affected_plans(affected_plans::Vector{String})

Print a list of the plans affected by a change and return how many plans were affected.